refactor(HomeAdapter): 네이밍 변경
